+++ docs/changelog.rst
@@ -44,6 +44,8 @@ Additions to the :doc:`supporting modules <supporting-modules>`.
     * :py:func:`khoros.utils.tests.test_core_utils.test_html_entity_convert`
     * :py:func:`khoros.utils.tests.test_core_utils.test_decode_binary`
     * :py:func:`khoros.utils.tests.test_core_utils.test_base64_conversions`
+* Added the :py:mod:`khoros.utils.tests.test_exceptions` module with the
+  :py:func:`khoros.utils.tests.test_exceptions.test_raising_exceptions` function.
 
 Changed
 =======